+++ CHANGES.md
@@ -9,6 +9,7 @@
 	- `;`: Jump to the next entry matching the most recent find command (repeatable)
 	- `,`: Jump to the previous entry matching the most recent find command (repeatable)
 - Fixed hanging when opening files when your default editor takes awhile to open (looking at you, Visual Studio)
+- Implemented delete selected action, `x` ([#8](https://github.com/LPGhatguy/magic-school-bus/issues/8))
 
 ## 0.3.1
 - Added `g` (top of list) and `G` bottom of list motions ([#5](https://github.com/LPGhatguy/magic-school-bus/issues/5))